Don't count the blocks above the water level
This adapts the bathymetry to the water level so that
the blocks which stick out of the water are not counted
negatively (as they were before). This also means we can
get rid of the maximum information.

This is a major breaking change. Areas computed using the old
method will not be equivalent to areas computed using the new
method